move util package to internal/utiltest
