+++ internal/utiltest/test_common.go
@@ -1,4 +1,4 @@
-package util
+package utiltest
 
 /*
  Copyright 2022 Crunchy Data Solutions, Inc.
